refactor resize-on-fontsize-change to hook
