+++ src/common/components/mock-components/front-text-components/heading1-text-shape.tsx
@@ -1,4 +1,4 @@
-import { forwardRef, useEffect, useRef } from 'react';
+import { forwardRef } from 'react';
 import { Group, Text } from 'react-konva';
 import { ShapeProps } from '../shape.model';
 import { ShapeSizeRestrictions, ShapeType } from '@/core/model';
@@ -6,8 +6,7 @@ import { fitSizeToShapeSizeRestrictions } from '@/common/utils/shapes/shape-rest
 import { useShapeProps } from '../../shapes/use-shape-props.hook';
 import { BASIC_SHAPE } from '../front-components/shape.const';
 import { useGroupShapeProps } from '../mock-components.utils';
-import { calcTextDimensions } from '@/common/utils/calc-text-dimensions';
-import { useCanvasContext } from '@/core/providers';
+import { useResizeOnFontSizeChange } from './front-text-hooks/resize-fontsize-change.hook';
 
 const heading1SizeRestrictions: ShapeSizeRestrictions = {
   minWidth: 40,
@@ -45,8 +44,6 @@ export const Heading1Shape = forwardRef<any, ShapeProps>((props, ref) => {
   const { textColor, textDecoration, fontStyle, fontVariant, fontSize } =
     useShapeProps(otherProps, BASIC_SHAPE);
 
-  const previousFontSize = useRef(fontSize);
-
   const commonGroupProps = useGroupShapeProps(
     props,
     restrictedSize,
@@ -54,26 +51,7 @@ export const Heading1Shape = forwardRef<any, ShapeProps>((props, ref) => {
     ref
   );
 
-  const { updateShapeSizeAndPosition, stageRef } = useCanvasContext();
-  const konvaLayer = stageRef.current?.getLayers()[0];
-
-  useEffect(() => {
-    if (previousFontSize.current != fontSize) {
-      previousFontSize.current = fontSize;
-      const { width, height } = calcTextDimensions(
-        text,
-        fontSize,
-        fontVariant,
-        konvaLayer
-      );
-      updateShapeSizeAndPosition(
-        id,
-        { x, y },
-        { width: width * 1.1, height },
-        false
-      );
-    }
-  }, [fontSize]);
+  useResizeOnFontSizeChange(id, { x, y }, text, fontSize, fontVariant);
 
   return (
     <Group {...commonGroupProps} {...shapeProps}>
